requirements.txt
pandas
numpy
sentence-transformers
scikit-learn
hdbscan
umap-learn
kagglehub
scikit-learn
